+++ src/memory.ts
@@ -131,3 +131,20 @@ When('I save screenshot as {string}', async function(key: string) {
     const screenshot = await browser.takeScreenshot();
     memory.setValue(key, screenshot);
 });
+
+/**
+ * Save css property of element to memory
+ * @param {string} property - css property to store
+ * @param {string} alias - element to get value
+ * @param {string} key - key to store value
+ * @example I save 'color' css property of 'Checkbox' as 'checkboxColor'
+ * @example I save '$propertyName' property of 'Checkbox' as 'checkboxColor'
+ */
+When('I save {string} css property of {string} as {string}', async function (property, alias, key) {
+    const element = await getElement(alias) as ElementAsync;
+    const propertyName = await getValue(property);
+    const value = await browser.execute(function (element: Element, propertyName: string) {
+        return getComputedStyle(element).getPropertyValue(propertyName)
+    }, element as any, propertyName);
+    memory.setValue(key, value);
+});