+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Current
 - added I click back/forward button step
+- added I save css property step
+- added I expect css property step
+- fixed after issue if browser not started
 
 ## 0.0.11
 - added I scroll to {string} step